Adds general configuration concept
Amay will have at least three levels of configuration files. This commit
actually starts the first steps to implement the top-level
configuration. This top-level will have generic configuration for the
customer to start VM has they see fit.

This is a simple proof of concept and the actual structure is in no way
representative of the final one. In fact, the `Domain` won't be at
top-level but instead at the domain-level configuration, per repo basis.